+++ config/__init__.py
@@ -1,0 +1,15 @@
+# ui/config/__init__.py
+
+# Optional: Import specific functions/classes from modules within the package
+# from .layouts import layout
+# from .callbacks import register_callbacks
+
+# You can define variables or perform other initialization actions here if needed
+# For example:
+# some_variable = 'value'
+
+# If your package structure is more complex, you might organize imports like this:
+# from .module1 import function1
+# from .module2 import Class2
+
+# Note: Importing specific components ensures cleaner namespace management
