+++ data/__init__.py
@@ -1,7 +1,7 @@
-# ui/__init__.py
+# ui/data/__init__.py
 
 # Optional: Import specific functions/classes from modules within the package
-from .layout_reference import create_layout
+# from .config.db_config import DB_SETTINGS
 
 # You can define variables or perform other initialization actions here if needed
 # For example:
