+++ data/db_import_prep.py
@@ -11,12 +11,14 @@ Step 4: Rename "Unit" to "unit" and "Test Standard" to "test_standard"
 # Load the CSV file
 df = pd.read_csv('parsed.csv')
 
+
 # Step 1: Split 'Value' column into 'value_double' and 'value_text'
 def identify_type(value):
     try:
         return float(value)
     except ValueError:
         return value
+
 
 df['identified_value'] = df['Value'].apply(identify_type)
 
@@ -31,11 +33,13 @@ property_columns = ['Rheological properties', 'Mechanical properties',
                     'Thermal properties', 'Electrical properties',
                     'Other properties']
 
+
 def determine_property_type_id(row):
     for i, col in enumerate(property_columns):
         if pd.notnull(row[col]):
             return i + 1
     return np.nan
+
 
 df['property_type_id'] = df.apply(determine_property_type_id, axis=1)
