--- server/db_formation/parsed.csv
@@ -1,35 +1,0 @@
-Rheological properties,Value,Unit,Test Standard,Mechanical properties,Thermal properties,Electrical properties,Other properties
-"Melt volume-flow rate, MVR",7,cm³/10min,ISO 1133,,,,
-Temperature,190,°C,ISO 1133,,,,
-Load,2.16,kg,ISO 1133,,,,
-,2000,MPa,ISO 527-1/-2,Tensile modulus,,,
-,50,MPa,ISO 527-1/-2,Yield stress,,,
-,12,%,ISO 527-1/-2,Yield strain,,,
-,45,%,ISO 527-1/-2,Nominal strain at break,,,
-,1550,MPa,ISO 899-1,"Tensile creep modulus, 1h",,,
-,1050,MPa,ISO 899-1,"Tensile creep modulus, 1000h",,,
-,N,kJ/m²,ISO 179/1eU,"Charpy impact strength, +23°C",,,
-,230,kJ/m²,ISO 179/1eU,"Charpy impact strength, -30°C",,,
-,11,kJ/m²,ISO 179/1eA,"Charpy notched impact strength, +23°C",,,
-,6.5,kJ/m²,ISO 179/1eA,"Charpy notched impact strength, -30°C",,,
-,167,°C,ISO 11357-1/-3,,"Melting temperature, 10°C/min",,
-,85,°C,ISO 75-1/-2,,"Temp. of deflection under load, 1.80 MPa",,
-,150,°C,ISO 75-1/-2,,"Temp. of deflection under load, 0.45 MPa",,
-,140,°C,ISO 306,,"Vicat softening temperature, 50°C/h 50N",,
-,120,E-6/K,ISO 11359-1/-2,,"Coeff. of linear therm. expansion, parallel",,
-,HB,class,IEC 60695-11-10,,Burning behavior at 1.5 mm nominal thickness,,
-,1.6,mm,IEC 60695-11-10,,Thickness tested (1.5),,
-,Yes,-,-,,Yellow Card available,,
-,HB,class,IEC 60695-11-10,,Burning behavior at thickness h,,
-,0.8,mm,IEC 60695-11-10,,Thickness tested (h),,
-,4,-,IEC 62631-2-1,,,"Relative permittivity, 100Hz",
-,4,-,IEC 62631-2-1,,,"Relative permittivity, 1MHz",
-,100,E-4,IEC 62631-2-1,,,"Dissipation factor, 100Hz",
-,140,E-4,IEC 62631-2-1,,,"Dissipation factor, 1MHz",
-,1E12,Ohm*m,IEC 62631-3-1,,,Volume resistivity,
-,1E14,Ohm,IEC 62631-3-2,,,Surface resistivity,
-,40,kV/mm,IEC 60243-1,,,Electric strength,
-,600,-,IEC 60112,,,Comparative tracking index,
-,0.8,%,Sim. to ISO 62,,,,Water absorption
-,0.2,%,Sim. to ISO 62,,,,Humidity absorption
-,1370,kg/m³,ISO 1183,,,,Density
