--- server/db_formation/table_1.csv
@@ -1,4 +1,0 @@
-Rheological properties,Value,Unit,Test Standard
-"Melt volume-flow rate, MVR",7,cm³/10min,ISO 1133
-Temperature,190,°C,ISO 1133
-Load,2.16,kg,ISO 1133
