--- server/db_formation/table_3.csv
@@ -1,11 +1,0 @@
-Thermal properties,Value,Unit,Test Standard
-"Melting temperature, 10°C/min",167,°C,ISO 11357-1/-3
-"Temp. of deflection under load, 1.80 MPa",85,°C,ISO 75-1/-2
-"Temp. of deflection under load, 0.45 MPa",150,°C,ISO 75-1/-2
-"Vicat softening temperature, 50°C/h 50N",140,°C,ISO 306
-"Coeff. of linear therm. expansion, parallel",120,E-6/K,ISO 11359-1/-2
-Burning behavior at 1.5 mm nominal thickness,HB,class,IEC 60695-11-10
-Thickness tested (1.5),1.6,mm,IEC 60695-11-10
-Yellow Card available,Yes,-,-
-Burning behavior at thickness h,HB,class,IEC 60695-11-10
-Thickness tested (h),0.8,mm,IEC 60695-11-10
