--- server/db_formation/table_5.csv
@@ -1,4 +1,0 @@
-Other properties,Value,Unit,Test Standard
-Water absorption,0.8,%,Sim. to ISO 62
-Humidity absorption,0.2,%,Sim. to ISO 62
-Density,1370,kg/m³,ISO 1183
